Migrate helper functions to utils.h, cleaner code now!
